pyproject.toml
[project]
name = "ehh"
version = "0.0.1"
authors = [{ name = "Fey Xie", email = "feyxiexzf@gmail.com" }]
description = "Utility for many homework platforms"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
dependencies = [
    "munch",
    "openai",
    "json5",
    "pyyaml",
    "httpx",
    "httpx[socks]",
    "platformdirs",
    "prompt_toolkit>=3",
    "rich>=14",
    "beautifulsoup4",
    "pyperclip",
]

[project.optional-dependencies]
tg-bot = ["python-telegram-bot"]
transcription = ["openai-whisper", "torch", "torchvision"]

[project.urls]
Homepage = "https://github.com/Ujhhgtg/extensible-homework-helper"
Issues = "https://github.com/Ujhhgtg/extensible-homework-helper/issues"
